se borro el directorio users y se corrigio el nombre en el readme.md
