+++ .babelrc
@@ -1,7 +1,8 @@
 {
   "presets": [
     "es2015",
-    "react"
+    "react",
+    "stage-0"
   ]
 
 }
