+++ app/common/redux/reducers/root-reducer.js
@@ -1,0 +1,5 @@
+import { combineReducers } from "redux";
+
+const rootReducer = combineReducers({});
+
+export default rootReducer;
