+++ app/common/redux/store.js
@@ -1,0 +1,6 @@
+import { createStore } from 'redux';
+import rootReducer from "./reducers/root-reducer";
+
+const store = createStore(rootReducer);
+
+export default store;
