+++ app/home/test.css
@@ -1,0 +1,3 @@
+.test {
+    background-color: deeppink;
+}
