+++ app/root.js
@@ -6,7 +6,7 @@ import App from './app';
 import store from './common/redux/store';
 
 ReactDom.render(
-    <Provider store={store}>
+    <Provider store={ store } >
         <App/>
     </Provider>,
     document.getElementById('app')
