+++ package.json
@@ -4,8 +4,9 @@
   "description": "",
   "main": "server.dev.js",
   "scripts": {
-    "dev": "webpack --config ./config/webpack.config.dev.js -p",
-    "prod": "webpack --config ./config/webpack.config.js -p"
+    "start:dev": "node_modules/webpack-dev-server/bin/webpack-dev-server.js --hot --content-base /",
+    "start:prod": "npm run build:app && node_modules/nodemon/bin/nodemon.js   -r dotenv/config -- node_modules/babel-cli/bin/babel-node.js dist/server.js dotenv_config_path=./config/.env",
+    "build:app": "webpack --config ./config/webpack.config.js -p"
   },
   "author": "Mauricio Irace",
   "license": "ISC",
@@ -17,25 +18,31 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.5",
+    "dotenv": "^4.0.0",
+    "express": "^4.15.4",
     "history": "^4.7.2",
-    "postcss-loader": "^2.0.6",
+    "morgan": "^1.8.2",
+    "nodemon": "^1.11.0",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-redux": "^5.0.6",
     "react-router": "^4.1.2",
     "react-router-dom": "^4.1.2",
+    "react-router-redux": "^4.0.8",
     "redux": "^3.7.2",
     "sass-loader": "^6.0.6",
     "socket.io": "^2.0.3",
-    "style-loader": "^0.18.2",
-    "svg-react-loader": "^0.4.4"
+    "style-loader": "^0.18.2"
   },
   "devDependencies": {
+    "babel-cli": "^6.26.0",
     "babel-loader": "^7.1.1",
+    "babel-preset-stage-0": "^6.24.1",
     "http-server": "^0.10.0",
+    "image-webpack-loader": "^3.4.0",
     "node-sass": "^4.5.3",
     "webpack": "^3.5.5",
-    "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.18.2"
+    "webpack-dev-server": "^2.7.1",
+    "webpack-dotenv-plugin": "^2.0.2"
   }
 }
