--- src/common/redux/reducers/root-reducer.js
@@ -1,6 +1,0 @@
-import userReducer from '../../../users/redux/user-reducer';
-import {combineReducers} from "redux";
-
-const rootReducer = combineReducers(userReducer());
-
-export default rootReducer;
