--- src/home/test.css
@@ -1,3 +1,0 @@
-.test{
-    color: deeppink;
-}
