--- src/users/redux/map-dispatch.js
@@ -1,9 +1,0 @@
-import setUser from './set-user';
-
-const mapDispatchToProps = (dispatch, props) => {
-    return {
-        setUser: (users) => dispatch(setUser(users))
-    }
-};
-
-export default mapDispatchToProps;
