--- src/users/redux/map-state.js
@@ -1,6 +1,0 @@
-const mapStateToProps = (state, props) => {
-    return {
-        users:state.users
-    };
-};
-export default mapStateToProps;
