--- src/users/redux/set-user.js
@@ -1,7 +1,0 @@
-const setUser = (users) => {
-    return {
-        type:'SET_USER',
-        users:users
-    };
-};
-export default setUser;
